Add init and system tray modules
